requirements.txt
fastapi[standard]
uvicorn[standard]
transformers
torch
Pillow